feat: specify public URL in .env file
the public URL value used in the webpack configuration originally
had a hardcoded value. this commit now makes the user specify that URL
in the .env file located at the project root. since the .env file is
only used during development or build time, the 'dotenv' package is a
development dependency